pyproject.toml
[project]
name = "final-lab-django-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.1.5",
    "djangorestframework==3.15.2",
    "djoser==2.3.1",
    "django-filter==24.3",
]
